requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.2
pip==21.1
pytz==2021.1
requests==2.25.1
setuptools==56.0.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.4
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
idna==3.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
oauthlib==3.1.0
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4
